backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.6.4
